refactor: move signup form to separate file
