+++ packages/web/src/pages/signup/index.tsx
@@ -1,44 +1,14 @@
-import clsx from 'clsx'
 import Head from 'next/head'
 import Image from 'next/image'
-import { useState } from 'react'
-import { useForm } from 'react-hook-form'
 
-import { StepOne, stepOneFields } from '@/screens/signup/step-1'
-import { StepTwo, stepTwoFields } from '@/screens/signup/step-2'
-import { StepThree, stepThreeFields } from '@/screens/signup/step-3'
-import { Steps } from '@/screens/signup/steps'
-import { Button } from '@/shared/components/ui/button'
+import { SignUpForm } from '@/screens/signup/form'
 import { Link } from '@/shared/components/ui/link'
 import SocialMediaLinks from '@/shared/components/ui/social-media-links'
-import { useYupValidationResolver } from '@/shared/hooks/useYupValidationResolver'
-import { SignUpFormFields } from '@/shared/types/signup-forms'
 import styles from '@/styles/classes'
-
-import { validationSchema } from './validators'
 
 import Logo from '@public/img/logo.svg'
 
 export default function SignUp() {
-  const resolver = useYupValidationResolver(validationSchema)
-  const form = useForm<SignUpFormFields>({ resolver, mode: 'onChange' })
-  const { handleSubmit } = form
-
-  const [currentStep, setCurrentStep] = useState(2)
-
-  const handleStep = async (step: number) => {
-    const steps = [stepOneFields, stepTwoFields, stepThreeFields]
-    const validStep = await form.trigger(steps[step])
-
-    if (validStep) {
-      setCurrentStep(step + 1)
-    }
-  }
-
-  const handleSendForm = (data: SignUpFormFields) => {
-    console.log(data)
-  }
-
   return (
     <main className="flex justify-center items-center">
       <Head>
@@ -65,79 +35,7 @@ export default function SignUp() {
           </div>
         </section>
 
-        <section className="md:w-full lg:w-10/12">
-          <Steps currentStep={currentStep} />
-
-          <div className="w-full bg-black-light rounded-3xl p-8">
-            <div className="flex gap-1 md:gap-1.5">
-              <h1 className="font-bold text-lg sm:text-2xl lg:text-3xl align-bottom">
-                Crie sua conta na
-              </h1>
-              <Image src={Logo} alt="HUBI" className="w-12 sm:w-16 lg:w-20" />
-            </div>
-            <p className="text-gray text-xs lg:text-sm mt-2 mb-10 text-justify">
-              O seu cadastro irá passar por um processo de verificação e pode
-              demorar para ser analisado.
-            </p>
-
-            <form onSubmit={handleSubmit(handleSendForm)} noValidate>
-              <div>
-                <StepOne form={form} hidden={currentStep !== 0} />
-                <StepTwo form={form} hidden={currentStep !== 1} />
-                <StepThree form={form} hidden={currentStep !== 2} />
-              </div>
-
-              <div className="font-light text-gray text-xs italic mt-2">
-                Campos com <span className="text-yellow">*</span> são
-                obrigatórios.
-              </div>
-
-              <div
-                className={clsx('text-gray text-xs italic mt-2', {
-                  hidden: currentStep !== 2,
-                })}
-              >
-                As redes sociais <span className="text-yellow">Twitter</span> e{' '}
-                <span className="text-yellow">Instagram</span> não podem estar
-                privadas.
-              </div>
-
-              <div className="flex gap-5">
-                {currentStep !== 0 && (
-                  <Button
-                    color="secondary"
-                    label="Voltar"
-                    className="w-full my-5"
-                    type="button"
-                    onClick={() => setCurrentStep(currentStep - 1)}
-                  />
-                )}
-                {currentStep !== 2 && (
-                  <Button
-                    color="primary"
-                    label="Continuar"
-                    className="w-full my-5"
-                    type="button"
-                    onClick={() => handleStep(currentStep)}
-                  />
-                )}
-                {currentStep === 2 && (
-                  <Button
-                    color="primary"
-                    label="Salvar"
-                    className="w-full my-5"
-                    type="submit"
-                  />
-                )}
-              </div>
-            </form>
-
-            <div className="text-center text-xs md:text-sm">
-              Já possui uma conta?{' '}
-              <Link text="Entre aqui &gt;" link="" className="font-medium" />
-            </div>
-          </div>
-        </section>
+        <SignUpForm />
 
         <section className="flex flex-col items-center mt-2 w-full lg:items-center lg:justify-center">
           <div className="lg:items-start">
